+++ components/common/src/cli.rs
@@ -20,12 +20,11 @@
 use crate::types::ListenCtlAddr;
 use clap::ArgMatches;
 
-#[cfg(unix)]
-use habitat_core::os::process::ShutdownSignal;
 use habitat_core::{env as henv,
                    fs::{cache_key_path,
                         CACHE_KEY_PATH},
-                   os::process::ShutdownTimeout};
+                   os::process::{ShutdownSignal,
+                                 ShutdownTimeout}};
 use std::path::PathBuf;
 
 pub const GOSSIP_DEFAULT_IP: &str = "0.0.0.0";
@@ -50,7 +49,10 @@ pub const PACKAGE_TARGET_ENVVAR: &str = "HAB_PACKAGE_TARGET";
 lazy_static! {
     pub static ref SHUTDOWN_TIMEOUT_DEFAULT: String = ShutdownTimeout::default().to_string();
 }
-#[cfg(unix)]
+
+// We allow this on Windows as well as Unix, even though we don't use
+// ShutdownSignal on Windows, because we want to allow Windows CLI
+// users to interact with Unix Supervisors.
 lazy_static! {
     pub static ref SHUTDOWN_SIGNAL_DEFAULT: String = ShutdownSignal::default().to_string();
 }
