+++ components/core/src/os/process/unix.rs
@@ -12,39 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::Signal;
 use crate::error::{Error,
                    Result};
 use libc::{self,
            pid_t};
 use std::{ffi::OsString,
-          fmt,
           io,
           os::unix::process::CommandExt,
           path::PathBuf,
-          process::Command,
-          result,
-          str::FromStr};
+          process::Command};
 
 pub type Pid = libc::pid_t;
 pub(crate) type SignalCode = libc::c_int;
-
-#[allow(non_snake_case)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum Signal {
-    INT,
-    ILL,
-    ABRT,
-    FPE,
-    KILL,
-    SEGV,
-    TERM,
-    HUP,
-    QUIT,
-    ALRM,
-    USR1,
-    USR2,
-    CHLD,
-}
 
 pub fn become_command(command: PathBuf, args: &[OsString]) -> Result<()> {
     become_exec_command(command, args)
@@ -76,6 +56,10 @@ pub fn signal(pid: Pid, signal: Signal) -> Result<()> {
     }
 }
 
+// This only makes sense on Unix platforms, because not all of these
+// symbols are actually defined on Windows. Also, this is only used
+// for actually sending the given signal to a process, which only
+// happens on Unix platforms anyway.
 impl From<Signal> for SignalCode {
     fn from(value: Signal) -> SignalCode {
         match value {
@@ -108,117 +92,4 @@ fn become_exec_command(command: PathBuf, args: &[OsString]) -> Result<()> {
     // The only possible return for the above function is an `Error` so return it, meaning that we
     // failed to exec to our target program
     Err(error_if_failed.into())
-}
-
-impl FromStr for Signal {
-    type Err = Error;
-
-    fn from_str(s: &str) -> result::Result<Self, Self::Err> {
-        let signal = match s {
-            "HUP" => Signal::HUP,
-            "INT" => Signal::INT,
-            "QUIT" => Signal::QUIT,
-            "ILL" => Signal::ILL,
-            "ABRT" => Signal::ABRT,
-            "FPE" => Signal::FPE,
-            "KILL" => Signal::KILL,
-            "USR1" => Signal::USR1,
-            "SEGV" => Signal::SEGV,
-            "USR2" => Signal::USR2,
-            "ALRM" => Signal::ALRM,
-            "TERM" => Signal::TERM,
-            "CHLD" => Signal::CHLD,
-            _ => return Err(Error::ParseSignalError(s.to_string())),
-        };
-        Ok(signal)
-    }
-}
-
-impl fmt::Display for Signal {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let s = match self {
-            Signal::HUP => "HUP",
-            Signal::INT => "INT",
-            Signal::QUIT => "QUIT",
-            Signal::ILL => "ILL",
-            Signal::ABRT => "ABRT",
-            Signal::FPE => "FPE",
-            Signal::KILL => "KILL",
-            Signal::USR1 => "USR1",
-            Signal::SEGV => "SEGV",
-            Signal::USR2 => "USR2",
-            Signal::ALRM => "ALRM",
-            Signal::TERM => "TERM",
-            Signal::CHLD => "CHLD",
-        };
-        write!(f, "{}", s)
-    }
-}
-
-/// Encapsulates logic for defining the default shutdown signal we
-/// send services, and handles translation from external types at the
-/// edges of our system.
-#[derive(Debug, Clone)]
-pub struct ShutdownSignal(Signal);
-
-impl Default for ShutdownSignal {
-    /// Unless otherwise specified, the Supervisor will shut down
-    /// services by sending the `TERM` signal.
-    fn default() -> Self { Signal::TERM.into() }
-}
-
-impl FromStr for ShutdownSignal {
-    type Err = Error;
-
-    fn from_str(s: &str) -> result::Result<Self, Self::Err> { Ok(ShutdownSignal(s.parse()?)) }
-}
-
-impl fmt::Display for ShutdownSignal {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result { write!(f, "{}", self.0) }
-}
-
-impl From<Signal> for ShutdownSignal {
-    fn from(signal: Signal) -> Self { ShutdownSignal(signal) }
-}
-
-impl From<ShutdownSignal> for Signal {
-    fn from(shutdown_signal: ShutdownSignal) -> Self { shutdown_signal.0 }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn signal_names_are_only_accepted_as_uppercase() {
-        assert_eq!(Signal::HUP, "HUP".parse().unwrap());
-        assert!("hup".parse::<Signal>().is_err());
-    }
-
-    #[test]
-    fn signals_can_render_as_strings() {
-        assert_eq!("HUP", Signal::HUP.to_string());
-    }
-
-    #[test]
-    fn signals_can_round_trip_through_parsing() {
-        for signal in &[Signal::HUP,
-                        Signal::INT,
-                        Signal::QUIT,
-                        Signal::ABRT,
-                        Signal::FPE,
-                        Signal::KILL,
-                        Signal::USR1,
-                        Signal::SEGV,
-                        Signal::USR2,
-                        Signal::ALRM,
-                        Signal::TERM,
-                        Signal::CHLD]
-        {
-            assert_eq!(*signal,
-                       signal.to_string()
-                             .parse::<Signal>()
-                             .expect("Couldn't parse back into a Signal!"));
-        }
-    }
 }
